requirements.txt
# Requirements for Medical Report Analyzer Backend

# Core Flask Backend (Option 1)
flask==3.0.0
flask-cors==4.0.0

# Core FastAPI Backend (Option 2)
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# File Processing Libraries
pillow==10.1.0          # Image processing
PyPDF2==3.0.1           # PDF text extraction
python-docx==1.1.0      # Word document processing

# Machine Learning (Add based on your needs)
# Uncomment the ones you need:

# scikit-learn==1.3.2   # For traditional ML models
# numpy==1.26.2         # Numerical computing
# pandas==2.1.4         # Data manipulation

# tensorflow==2.15.0    # For deep learning (if using Keras/TF)
# torch==2.1.2          # For PyTorch models
# torchvision==0.16.2   # For PyTorch image models

# joblib==1.3.2         # For model serialization

# Additional useful libraries
# opencv-python==4.8.1.78  # Advanced image processing
# pytesseract==0.3.10      # OCR for scanned documents
